Add dummy gnuplot impl for js
